+++ NEWS.md
@@ -1,5 +1,8 @@
 DataMan 0.5.1
 -------------------------------------------------------------
+NEW FEATURES
+* compPlot to compare different model predictions in a plot.
+
 BUG FIXES AND MINOR IMPROVEMENTS
 * The axis title of 3d plot can now be set correctly with latest version of `plotly` on github. (#31)